Rename fake-plugins to just fake
