+++ test/fake/fake-currency-info.js
