+++ test/fake/fake-currency-plugin.js
